+++ sdk/core/azure-core/test/ut/policy.cpp
@@ -59,12 +59,8 @@ struct TestContextTreeIntegrity : public Azure::Core::Http::Policies::HttpPolicy
       Azure::Core::Http::Policies::NextHttpPolicy nextHttpPolicy,
       Azure::Core::Context const& ctx) const override
   {
-    EXPECT_TRUE(ctx.HasKey("TheKey"));
-    if (ctx.HasKey("TheKey"))
-    {
-      auto value = ctx.Get<std::string>("TheKey");
-      EXPECT_EQ("TheValue", value);
-    }
+    auto value = ctx.Get<std::string>("TheKey");
+    EXPECT_EQ("TheValue", value);
     return nextHttpPolicy.Send(request, ctx);
   }
 };